Add HTML for reset button and displaying game text
The reset button will be used to restart the game at any point. It will be
particularly useful when the game is over and the user wants to play another
round.

In addition, game-text will be used to initially display how to make a move, and
when the game is over the results of the game.